feat: Implement recommendation engine with modular architecture
- Added Formatter interface and TextFormatter implementation for formatting recommendations.
- Refactored RecommendationEngine to utilize Analyzer, Scorer, Selector, Validator, and Formatter interfaces.
- Introduced Scorer interface with WeightedScorer implementation for calculating number priorities.
- Created Selector interface with ZoneBasedSelector for selecting numbers based on zones.
- Defined RecommendationConfig and StatisticalAnalysis types for configuration and analysis results.
- Implemented CompositeValidator for validating combinations with multiple validation strategies.
- Added specific validators for zone distribution, odd/even ratio, total sum, average value, smoothness, and close number pairs.
- Removed redundant methods and streamlined recommendation generation process.
